Move serializer components into app directory
Move all serializers that are used in the example into a separate app
directory in the project root and modify the corresponding import
statements. Given that there are only two options for serialization
available (JSON file & SQLite) and the number of files is small, go
without dedicated subdirectories/packages.
