+++ spiff_example/misc/custom_start_event.py
@@ -18,7 +18,7 @@ from SpiffWorkflow.bpmn.parser.util import full_tag
 from SpiffWorkflow.bpmn.serializer.default import EventConverter
 from SpiffWorkflow.spiff.serializer.task_spec import SpiffBpmnTaskConverter
 
-from ..serializer.file import FileSerializer
+from app.serializer import FileSerializer
 from ..engine import BpmnEngine
 from .curses_handlers import UserTaskHandler, ManualTaskHandler
