+++ spiff_example/misc/threaded_service_task.py
@@ -13,7 +13,7 @@ from SpiffWorkflow.spiff.parser.task_spec import ServiceTaskParser
 from SpiffWorkflow.bpmn.parser.util import full_tag
 from SpiffWorkflow.bpmn.exceptions import WorkflowTaskException
 
-from ..serializer.file import FileSerializer
+from app.serializer import FileSerializer
 from ..engine import BpmnEngine, Instance
 from .curses_handlers import UserTaskHandler, ManualTaskHandler
