--- spiff_example/serializer/file/__init__.py
@@ -1,1 +1,0 @@
-from .serializer import FileSerializer
