+++ spiff_example/spiff/custom_exec.py
@@ -9,7 +9,7 @@ from SpiffWorkflow.bpmn.specs.mixins.none_task import NoneTask
 from SpiffWorkflow.bpmn.script_engine.python_environment import BasePythonScriptEngineEnvironment
 from SpiffWorkflow.util.deep_merge import DeepMerge
 
-from ..serializer.file import FileSerializer
+from app.serializer import FileSerializer
 from ..engine import BpmnEngine
 from .curses_handlers import UserTaskHandler, ManualTaskHandler
